requirements.txt
torch==1.0.0
torchvision==0.2.1
python-dateutil==2.7.5
easydict==1.9
scikit-image==0.14.1
runway-python
scipy==1.1.0
numpy==1.15.0
pyyaml==3.13
